+++ noodles-bcf/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+  * reader: Accept any `BinningIndex` to query.
+
 ### Fixed
 
   * Fixed documentation link in package manifest ([#31]).